Add files via upload
The Myntra Clone is a fully responsive frontend replica of the popular Indian fashion e-commerce website, Myntra. This project was developed as a hands-on exercise to strengthen my skills in modern web development using HTML, CSS, and JavaScript. It captures the core layout, design aesthetics, and basic interactivity of an e-commerce platform